Move menu state up into layout so that cursor component has access to it
