+++ src/components/Layout.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { Component } from "react"
 import styled from "styled-components"
 import PropTypes from "prop-types"
 import Helmet from "react-helmet"
@@ -26,18 +26,42 @@ const HideMenu = styled.div`
   background-color: var(--white);
 `
 
-const Layout = ({ children, location }) => (
-  <StaticQuery
-    query={graphql`
-      query SiteTitleQuery {
-        site {
-          siteMetadata {
-            title
-          }
-        }
-      }
-    `}
-    render={data => (
+class LayoutWithState extends Component {
+  state = {
+    menuActive: false,
+    color: "yellow",
+    colorIndex: 0,
+  }
+
+  toggleMenu = () => {
+    if (!this.state.menuActive) {
+      this.setNewMenuColor()
+    }
+    this.setState(prevProps => ({
+      menuActive: !prevProps.menuActive,
+    }))
+  }
+
+  setNewMenuColor() {
+    const colors = ["yellow", "blue", "green", "red"]
+    this.setState(prevState => ({
+      color: colors[prevState.colorIndex % 4],
+      colorIndex: prevState.colorIndex + 1,
+    }))
+  }
+
+  render() {
+    const { data, children, location } = this.props
+
+    const childrenWithProps = React.Children.map(
+      children,
+      child =>
+        child.type.displayName === "Cursor"
+          ? React.cloneElement(child, { menuActive: this.state.menuActive })
+          : child
+    )
+
+    return (
       <>
         <Helmet
           title={data.site.siteMetadata.title}
@@ -52,12 +76,34 @@ const Layout = ({ children, location }) => (
         <AppWrapper>
           <HideMenu />
           <div className="relative">
-            <Menu location={location} />
+            <Menu
+              location={location}
+              menuActive={this.state.menuActive}
+              toggleMenu={this.toggleMenu}
+              color={this.state.color}
+            />
           </div>
           <Header siteTitle={data.site.siteMetadata.title} />
-          <div>{children}</div>
+          <div>{childrenWithProps}</div>
         </AppWrapper>
       </>
+    )
+  }
+}
+
+const Layout = ({ children, location }) => (
+  <StaticQuery
+    query={graphql`
+      query SiteTitleQuery {
+        site {
+          siteMetadata {
+            title
+          }
+        }
+      }
+    `}
+    render={data => (
+      <LayoutWithState children={children} location={location} data={data} /> // eslint-disable-line
     )}
   />
 )
